+++ elfo-core/src/start.rs
@@ -13,9 +13,8 @@ use crate::{
     message,
     messages::{Ping, UpdateConfig},
     object::{Object, ObjectMeta},
-    tls,
+    scope::Scope,
     topology::Topology,
-    trace_id,
 };
 
 type Result<T, E = StartError> = std::result::Result<T, E>;
@@ -94,12 +93,12 @@ pub async fn do_start<F: Future>(
         group: "starter".into(),
         key: Some("_".into()), // Just like `Singleton`.
     };
-    let initial_trace_id = trace_id::generate();
-    tls::scope(Arc::new(meta), initial_trace_id, async move {
+    let scope = Scope::new(Arc::new(meta));
+    let f = async move {
         let ctx = Context::new(topology.book.clone(), dumper, Demux::default()).with_addr(addr);
         send_configs_to_entrypoints(&ctx, &topology).await?;
         start_entrypoints(&ctx, &topology).await?;
         Ok(f(ctx).await)
-    })
-    .await
+    };
+    scope.within(f).await
 }